Refactored BudgetsChart to be reusable and updated associated stylings
