+++ styles/components/budgets/BudgetsChart.module.css
@@ -1,0 +1,17 @@
+.chart {
+  height: 15.4375rem;
+  width: 15rem;
+  justify-self: center;
+  position: relative;
+}
+
+.chart-text {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+  align-items: center;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
