+++ styles/components/overview/BudgetsOverview.module.css
@@ -15,24 +15,6 @@
   gap: 1rem;
 }
 
-.chart {
-  height: 15.4375rem;
-  width: 15rem;
-  justify-self: center;
-  position: relative;
-}
-
-.chart-text {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-  align-items: center;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-
 .details {
   display: flex;
   flex-direction: column;
